move the check functions to a new `utils` module
