+++ python/healpix_geo/utils.py
@@ -1,0 +1,15 @@
+import numpy as np
+
+
+def _check_depth(depth):
+    ravel_depth = np.ravel(np.atleast_1d(depth))
+    if any(ravel_depth < 0) or any(ravel_depth > 29):
+        raise ValueError("Depth must be in the [0, 29] closed range")
+
+
+def _check_ipixels(data, depth):
+    npix = 12 * 4**depth
+    if (data >= npix).any() or (data < 0).any():
+        raise ValueError(
+            f"The input HEALPix cells contains value out of [0, {npix - 1}]"
+        )
